Todo 昨日するの部分を共通化
